valmis

+++ Moduuli3/Tehtävä 1.py
@@ -1,4 +1,4 @@
-kuha = int(input("Anna kuhan pituus: "))
+kuha = float(input("Anna kuhan pituus: "))
 if kuha>=37:
     print("Voit pitää kalan, hyvää ruokahalua!:)")
 else:
